+++ aospy/utils/__init__.py
@@ -1,6 +1,9 @@
 """Subpackage comprising various utility functions used elsewhere in aospy."""
 from . import io
+from . import longitude
+from .longitude import Longitude
 from . import times
 from . import vertcoord
 
-__all__ = ['io', 'times', 'vertcoord']
+
+__all__ = ['Longitude', 'io', 'longitude', 'times', 'vertcoord']